refactor: reduce file sizes and improve code organization
- Split useImageEditor.ts (707→508 lines): extracted zoom, export, and keyboard shortcuts into separate hooks (useZoom, useExport, useKeyboardShortcuts)
- Split types.ts (573→7 lines): reorganized types into separate files (fabric.types, config.types, state.types, hooks.types, component.types) for better modularity
- Split useCrop.ts (486→305 lines): extracted UI helpers and coordinate calculations into crop.ts utility file
- Removed console.warn/error from production code (4 instances)
- All files now adhere to 400-line limit; build and TypeScript checks pass successfully
