+++ src/core/hooks/useImageEditor.ts
@@ -10,13 +10,7 @@ import type {
   ImageEditorConfig,
   UseImageEditorReturn,
 } from "../types";
-import {
-  animateZoomTo,
-  clampZoom,
-  CleanupManager,
-  addTrackedEventListener,
-  dataURLToBlob,
-} from "../utils";
+import { CleanupManager } from "../utils";
 
 import { useHistory } from "./useHistory";
 import { useBlur } from "./useBlur";
@@ -24,6 +18,9 @@ import { useCrop } from "./useCrop";
 import { useDrawing } from "./useDrawing";
 import { useShapes } from "./useShapes";
 import { useSelection } from "./useSelection";
+import { useZoom } from "./useZoom";
+import { useExport } from "./useExport";
+import { useKeyboardShortcuts } from "./useKeyboardShortcuts";
 
 // ============================================================================
 // Constants
@@ -57,33 +54,8 @@ const DEFAULT_CONFIG: Required<ImageEditorConfig> = {
  *   return (
  *     <div>
  *       <canvas ref={editor.canvasRef} />
- *
  *       <button onClick={editor.blur.add}>Add Blur</button>
  *       <button onClick={editor.crop.start}>Crop</button>
- *       <button onClick={editor.drawing.toggle}>
- *         {editor.drawing.isActive ? 'Stop Drawing' : 'Draw'}
- *       </button>
- *       <button onClick={editor.shapes.addRectangle}>Rectangle</button>
- *       <button onClick={editor.shapes.addCircle}>Circle</button>
- *       <button onClick={editor.history.undo} disabled={!editor.history.canUndo}>
- *         Undo
- *       </button>
- *       <button onClick={editor.history.redo} disabled={!editor.history.canRedo}>
- *         Redo
- *       </button>
- *
- *       <input
- *         type="color"
- *         value={editor.style.color}
- *         onChange={(e) => editor.style.setColor(e.target.value)}
- *       />
- *
- *       <button onClick={async () => {
- *         const blob = await editor.exportToBlob();
- *         if (blob) saveImage(blob);
- *       }}>
- *         Save
- *       </button>
  *     </div>
  *   );
  * }
@@ -108,7 +80,6 @@ export function useImageEditor(
   const [currentStrokeWidth, setCurrentStrokeWidth] = useState(
     mergedConfig.defaultStrokeWidth
   );
-  const [zoom, setZoomLevel] = useState(1);
 
   // Keep refs in sync
   useEffect(() => {
@@ -130,7 +101,6 @@ export function useImageEditor(
     ) as fabric.Image;
 
     if (imageObj) {
-      // Keep image locked
       imageObj.set({
         selectable: false,
         evented: false,
@@ -201,6 +171,20 @@ export function useImageEditor(
     defaultColor: currentColor,
     defaultStrokeWidth: currentStrokeWidth,
     onShapeAdd: () => history.saveState(),
+  });
+
+  // Zoom hook
+  const zoomHook = useZoom(canvasInstanceRef);
+
+  // Export hook
+  const exportHook = useExport(canvasInstanceRef, originalImageRef);
+
+  // Keyboard shortcuts
+  useKeyboardShortcuts({
+    onUndo: history.undo,
+    onRedo: history.redo,
+    onDelete: selectionHook.deleteSelected,
+    canDelete: () => !!selectionHook.selectedObject,
   });
 
   // Load image
@@ -284,7 +268,6 @@ export function useImageEditor(
     canvasInstanceRef.current = fabricCanvas;
     setCanvas(fabricCanvas);
 
-    // Load image if URL provided
     if (mergedConfig.imageUrl) {
       loadImage(mergedConfig.imageUrl);
     }
@@ -306,7 +289,6 @@ export function useImageEditor(
   useEffect(() => {
     if (!canvasInstanceRef.current || !mergedConfig.imageUrl) return;
 
-    // Clear existing objects except background
     const currentCanvas = canvasInstanceRef.current;
     const objects = currentCanvas.getObjects();
     objects.forEach((obj) => {
@@ -314,12 +296,10 @@ export function useImageEditor(
     });
     currentCanvas.renderAll();
 
-    // Reset state
     setOriginalImage(null);
     originalImageRef.current = null;
     setHasImage(false);
 
-    // Load new image
     loadImage(mergedConfig.imageUrl);
   }, [mergedConfig.imageUrl, loadImage]);
 
@@ -328,9 +308,6 @@ export function useImageEditor(
     const currentCanvas = canvasInstanceRef.current;
     if (!currentCanvas) return;
 
-    const cleanup = cleanupRef.current;
-
-    // Selection events
     const onSelectionCreated = (e: FabricSelectionEvent) => {
       if (e.selected?.[0]) {
         (selectionHook as unknown as { _handleSelect: (obj: fabric.Object) => void })._handleSelect(e.selected[0]);
@@ -347,7 +324,6 @@ export function useImageEditor(
       selectionHook.clearSelection();
     };
 
-    // Path created (drawing)
     const onPathCreated = (e: FabricPathCreatedEvent) => {
       if (e.path) {
         const path = e.path as EditorFabricPath;
@@ -356,7 +332,6 @@ export function useImageEditor(
       }
     };
 
-    // Object modified
     const onObjectModified = () => {
       history.saveState();
     };
@@ -368,37 +343,6 @@ export function useImageEditor(
     currentCanvas.on("path:created", onPathCreated as any);
     currentCanvas.on("object:modified", onObjectModified);
 
-    // Keyboard shortcuts
-    const onKeyDown = (e: KeyboardEvent) => {
-      // Undo: Ctrl/Cmd + Z
-      if ((e.ctrlKey || e.metaKey) && e.key === "z" && !e.shiftKey) {
-        e.preventDefault();
-        history.undo();
-      }
-      // Redo: Ctrl/Cmd + Y or Ctrl/Cmd + Shift + Z
-      if (
-        ((e.ctrlKey || e.metaKey) && e.key === "y") ||
-        ((e.ctrlKey || e.metaKey) && e.shiftKey && e.key === "z")
-      ) {
-        e.preventDefault();
-        history.redo();
-      }
-      // Delete: Delete or Backspace
-      if (e.key === "Delete" || e.key === "Backspace") {
-        const activeElement = document.activeElement;
-        const isInputFocused =
-          activeElement?.tagName === "INPUT" ||
-          activeElement?.tagName === "TEXTAREA";
-
-        if (!isInputFocused && selectionHook.selectedObject) {
-          e.preventDefault();
-          selectionHook.deleteSelected();
-        }
-      }
-    };
-
-    addTrackedEventListener(window, "keydown", onKeyDown, cleanup);
-
     return () => {
       currentCanvas.off("selection:created", onSelectionCreated);
       currentCanvas.off("selection:updated", onSelectionUpdated);
@@ -406,7 +350,6 @@ export function useImageEditor(
       // eslint-disable-next-line @typescript-eslint/no-explicit-any
       currentCanvas.off("path:created", onPathCreated as any);
       currentCanvas.off("object:modified", onObjectModified);
-      cleanup.cleanup();
     };
   }, [history, selectionHook]);
 
@@ -426,13 +369,11 @@ export function useImageEditor(
     (color: string) => {
       setCurrentColor(color);
 
-      // Update selected object if any
       if (selectionHook.selectedObject) {
         selectionHook.setSelectedColor(color);
         history.saveState();
       }
 
-      // Update drawing brush
       drawing.setBrushColor(color);
     },
     [selectionHook, drawing, history]
@@ -442,131 +383,14 @@ export function useImageEditor(
     (width: number) => {
       setCurrentStrokeWidth(width);
 
-      // Update selected object if any
       if (selectionHook.selectedObject) {
         selectionHook.setSelectedStrokeWidth(width);
         history.saveState();
       }
 
-      // Update drawing brush
       drawing.setBrushWidth(width);
     },
     [selectionHook, drawing, history]
-  );
-
-  // Zoom handlers
-  const zoomIn = useCallback(
-    (factor: number = 1.2) => {
-      const currentCanvas = canvasInstanceRef.current;
-      if (!currentCanvas) return;
-
-      const newZoom = clampZoom(currentCanvas.getZoom() * factor);
-      animateZoomTo(currentCanvas, newZoom);
-      setZoomLevel(newZoom);
-    },
-    []
-  );
-
-  const zoomOut = useCallback(
-    (factor: number = 1.2) => {
-      const currentCanvas = canvasInstanceRef.current;
-      if (!currentCanvas) return;
-
-      const newZoom = clampZoom(currentCanvas.getZoom() / factor);
-      animateZoomTo(currentCanvas, newZoom);
-      setZoomLevel(newZoom);
-    },
-    []
-  );
-
-  const setZoom = useCallback((level: number) => {
-    const currentCanvas = canvasInstanceRef.current;
-    if (!currentCanvas) return;
-
-    const clampedZoom = clampZoom(level);
-    animateZoomTo(currentCanvas, clampedZoom);
-    setZoomLevel(clampedZoom);
-  }, []);
-
-  const resetZoom = useCallback(() => {
-    setZoom(1);
-  }, [setZoom]);
-
-  // Export handlers
-  const exportToBlob = useCallback(async (): Promise<Blob | null> => {
-    const currentCanvas = canvasInstanceRef.current;
-    const image = originalImageRef.current;
-    if (!currentCanvas || !image) return null;
-
-    try {
-      // Store viewport state
-      const currentVPT = currentCanvas.viewportTransform || [1, 0, 0, 1, 0, 0];
-      const currentZoom = currentCanvas.getZoom();
-
-      // Reset viewport
-      currentCanvas.setZoom(1);
-      currentCanvas.setViewportTransform([1, 0, 0, 1, 0, 0]);
-
-      // Get bounds
-      const imgLeft = image.left || 0;
-      const imgTop = image.top || 0;
-      const imgWidth = image.getScaledWidth();
-      const imgHeight = image.getScaledHeight();
-
-      let minX = imgLeft;
-      let minY = imgTop;
-      let maxX = imgLeft + imgWidth;
-      let maxY = imgTop + imgHeight;
-
-      // Include other objects
-      currentCanvas.getObjects().forEach((obj) => {
-        if (obj !== image) {
-          const bounds = obj.getBoundingRect();
-          minX = Math.min(minX, bounds.left);
-          minY = Math.min(minY, bounds.top);
-          maxX = Math.max(maxX, bounds.left + bounds.width);
-          maxY = Math.max(maxY, bounds.top + bounds.height);
-        }
-      });
-
-      // Clamp to canvas
-      const canvasWidth = currentCanvas.getWidth();
-      const canvasHeight = currentCanvas.getHeight();
-      minX = Math.max(0, minX);
-      minY = Math.max(0, minY);
-      maxX = Math.min(canvasWidth, maxX);
-      maxY = Math.min(canvasHeight, maxY);
-
-      const dataURL = currentCanvas.toDataURL({
-        format: "png",
-        quality: 1,
-        left: Math.round(minX),
-        top: Math.round(minY),
-        width: Math.round(maxX - minX),
-        height: Math.round(maxY - minY),
-        multiplier: 1,
-      });
-
-      // Restore viewport
-      currentCanvas.setZoom(currentZoom);
-      currentCanvas.setViewportTransform(currentVPT);
-      currentCanvas.renderAll();
-
-      return await dataURLToBlob(dataURL);
-    } catch (error) {
-      console.error("Export error:", error);
-      return null;
-    }
-  }, []);
-
-  const exportToDataURL = useCallback(
-    (format: string = "png", quality: number = 1): string | null => {
-      const currentCanvas = canvasInstanceRef.current;
-      if (!currentCanvas) return null;
-
-      return currentCanvas.toDataURL({ format, quality });
-    },
-    []
   );
 
   // Dispose
@@ -600,7 +424,7 @@ export function useImageEditor(
       currentStrokeWidth,
       canUndo: history.canUndo,
       canRedo: history.canRedo,
-      zoom,
+      zoom: zoomHook.zoom,
     }),
     [
       hasImage,
@@ -612,21 +436,16 @@ export function useImageEditor(
       currentStrokeWidth,
       history.canUndo,
       history.canRedo,
-      zoom,
+      zoomHook.zoom,
     ]
   );
 
   return {
-    // Canvas
     canvas,
     canvasRef,
     originalImage,
     hasImage,
-
-    // State
     state,
-
-    // History
     history: {
       save: history.saveState,
       undo: history.undo,
@@ -635,39 +454,29 @@ export function useImageEditor(
       canRedo: history.canRedo,
       clear: history.clearHistory,
     },
-
-    // Blur
     blur: {
       add: blur.addBlur,
       remove: blur.removeBlur,
       clearAll: blur.clearAllBlur,
       activeRects: blur.activeBlurRects,
     },
-
-    // Crop
     crop: {
       start: crop.startCrop,
       apply: crop.applyCrop,
       cancel: crop.cancelCrop,
       isActive: crop.isCropping,
     },
-
-    // Drawing
     drawing: {
       enable: drawing.enableDrawing,
       disable: drawing.disableDrawing,
       toggle: drawing.toggleDrawing,
       isActive: drawing.isDrawing,
     },
-
-    // Shapes
     shapes: {
       addRectangle: shapes.addRectangle,
       addCircle: shapes.addCircle,
       add: shapes.addShape,
     },
-
-    // Selection
     selection: {
       enable: selectionHook.enableSelectMode,
       disable: selectionHook.disableSelectMode,
@@ -677,29 +486,21 @@ export function useImageEditor(
       clear: selectionHook.clearSelection,
       isActive: selectionHook.isSelectMode,
     },
-
-    // Style
     style: {
       color: currentColor,
       setColor,
       strokeWidth: currentStrokeWidth,
       setStrokeWidth,
     },
-
-    // Zoom
     zoom: {
-      level: zoom,
-      in: zoomIn,
-      out: zoomOut,
-      set: setZoom,
-      reset: resetZoom,
+      level: zoomHook.zoom,
+      in: zoomHook.zoomIn,
+      out: zoomHook.zoomOut,
+      set: zoomHook.setZoom,
+      reset: zoomHook.resetZoom,
     },
-
-    // Export
-    exportToBlob,
-    exportToDataURL,
-
-    // Lifecycle
+    exportToBlob: exportHook.exportToBlob,
+    exportToDataURL: exportHook.exportToDataURL,
     dispose,
   };
 }
